+++ pytest_wdl/executors/__init__.py
@@ -24,7 +24,7 @@ from pytest_wdl.utils import (
     ensure_path, safe_string, find_executable_path, find_in_classpath
 )
 
-from WDL import Document, Tree
+from WDL import Document, Error, Tree
 
 
 ENV_JAVA_HOME = "JAVA_HOME"
@@ -179,6 +179,7 @@ class Executor(metaclass=ABCMeta):
                     f"Expected and actual values differ in length for {name}: "
                     f"{len(expected_value)} != {len(actual_value)}"
                 )
+
             for i, (exp, act) in enumerate(zip(expected_value, actual_value)):
                 cls._compare_output_values(exp, act, f"{name}[{i}]")
         elif isinstance(expected_value, dict):
@@ -187,8 +188,14 @@ class Executor(metaclass=ABCMeta):
                     f"Expected and actual values differ in length for {name}: "
                     f"{len(expected_value)} != {len(actual_value)}"
                 )
+
             for key, exp in expected_value.items():
-                assert key in actual_value
+                if key not in actual_value:
+                    raise AssertionError(
+                        f"Key '{key}' is in the expected value but not in "
+                        f"the actual value: {expected_value} != {actual_value}"
+                    )
+
                 cls._compare_output_values(exp, actual_value[key], f"{name}.{key}")
         elif isinstance(expected_value, DataFile):
             # TODO: pass name
@@ -356,7 +363,14 @@ def get_target_name(
         return workflow_name, False
 
     if not wdl_doc and Tree:
-        wdl_doc = parse_wdl(wdl_path, **kwargs)
+        try:
+            wdl_doc = parse_wdl(wdl_path, **kwargs)
+        except Error.SyntaxError as err:
+            raise RuntimeError(
+                "There was an error parsing the WDL document to extract the target "
+                "workflow/task name. Please specify the 'workflow_name' or 'task_name' "
+                "parameter to workflow_runner()."
+            ) from err
 
     if wdl_doc:
         if wdl_doc.workflow:
@@ -379,6 +393,7 @@ def read_write_inputs(
     inputs_file: Optional[Union[str, Path]] = None,
     inputs_dict: Optional[dict] = None,
     inputs_formatter: Optional[InputsFormatter] = InputsFormatter.get_instance(),
+    write_formatted_inputs: bool = True,
     **kwargs
 ) -> Tuple[dict, Optional[Path]]:
     """
@@ -390,6 +405,7 @@ def read_write_inputs(
         inputs_file:
         inputs_dict:
         inputs_formatter:
+        write_formatted_inputs:
         kwargs:
 
     Returns:
@@ -405,13 +421,14 @@ def read_write_inputs(
                 return inputs_dict_from_file, inputs_file
 
     if inputs_dict:
-        if not inputs_file:
-            inputs_file = Path(tempfile.mkstemp(suffix=".json")[1])
-
         inputs_dict = inputs_formatter.format_inputs(inputs_dict, **kwargs)
 
-        with open(inputs_file, "wt") as out:
-            json.dump(inputs_dict, out, default=str)
+        if write_formatted_inputs:
+            if not inputs_file:
+                inputs_file = Path(tempfile.mkstemp(suffix=".json")[1])
+
+            with open(inputs_file, "wt") as out:
+                json.dump(inputs_dict, out, default=str)
 
         return inputs_dict, inputs_file