+++ tests/test_cromwell.py
@@ -22,8 +22,8 @@ import pytest
 
 from pytest_wdl.utils import ENV_PATH, ENV_CLASSPATH
 from pytest_wdl.executors import ENV_JAVA_HOME
-from pytest_wdl.executors.cromwell import (
-    ENV_CROMWELL_CONFIG, ENV_CROMWELL_JAR, CromwellExecutor
+from pytest_wdl.executors.cromwell_local import (
+    ENV_CROMWELL_CONFIG, ENV_CROMWELL_JAR,  CromwellLocalExecutor
 )
 from . import setenv, make_executable
 
@@ -41,7 +41,7 @@ def test_java_bin(user_config):
         make_executable(java)
 
         with setenv({ENV_JAVA_HOME: str(d)}):
-            assert CromwellExecutor(
+            assert  CromwellLocalExecutor(
                 [d], cromwell_jar_file=cromwell_jar_file
             ).java_bin == java
 
@@ -49,7 +49,7 @@ def test_java_bin(user_config):
             ENV_PATH: str(d / "bin"),
             ENV_JAVA_HOME: None
         }):
-            assert CromwellExecutor(
+            assert  CromwellLocalExecutor(
                 [d], cromwell_jar_file=cromwell_jar_file
             ).java_bin == java
 
@@ -58,13 +58,13 @@ def test_java_bin(user_config):
             ENV_JAVA_HOME: None
         }):
             with pytest.raises(FileNotFoundError):
-                assert CromwellExecutor(
+                assert  CromwellLocalExecutor(
                     [d], cromwell_jar_file=cromwell_jar_file
                 ).java_bin
 
     with setenv({ENV_JAVA_HOME: "foo"}):
         with pytest.raises(FileNotFoundError):
-            assert CromwellExecutor([d]).java_bin
+            assert  CromwellLocalExecutor([d]).java_bin
 
 
 def test_cromwell_config(user_config):
@@ -73,18 +73,21 @@ def test_cromwell_config(user_config):
     )
 
     with tempdir() as d:
-        assert CromwellExecutor(
-            [d], cromwell_jar_file=cromwell_jar_file
-        )._cromwell_config_file is None
+        executor = CromwellLocalExecutor([d], cromwell_jar_file=cromwell_jar_file)
+        assert executor._cromwell_args is None
+        assert executor.java_args is None
+
         config = d / "config"
         with setenv({ENV_CROMWELL_CONFIG: str(config)}):
             with pytest.raises(FileNotFoundError):
-                CromwellExecutor([d], cromwell_jar_file=cromwell_jar_file)
+                 CromwellLocalExecutor([d], cromwell_jar_file=cromwell_jar_file)
             with open(config, "wt") as out:
                 out.write("foo")
-            assert CromwellExecutor(
+            executor =  CromwellLocalExecutor(
                 [d], cromwell_jar_file=cromwell_jar_file
-            )._cromwell_config_file == config
+            )
+            assert executor._cromwell_args is None
+            assert executor.java_args == f"-Dconfig.file={config}"
 
 
 def test_java_args(user_config):
@@ -93,28 +96,28 @@ def test_java_args(user_config):
     )
 
     with tempdir() as d:
-        assert CromwellExecutor(
+        assert  CromwellLocalExecutor(
             [d], cromwell_jar_file=cromwell_jar_file
         ).java_args is None
 
         with pytest.raises(FileNotFoundError):
-            CromwellExecutor(
+             CromwellLocalExecutor(
                 [d],
-                cromwell_config_file=Path("foo"),
+                cromwell_configuration=Path("foo"),
                 cromwell_jar_file=cromwell_jar_file
             ).java_args
 
         config = d / "config"
         with pytest.raises(FileNotFoundError):
-            CromwellExecutor(
+             CromwellLocalExecutor(
                 [d],
-                cromwell_config_file=Path("foo"),
+                cromwell_configuration=Path("foo"),
                 cromwell_jar_file=cromwell_jar_file,
             ).java_args
         with open(config, "wt") as out:
             out.write("foo")
-        assert CromwellExecutor(
-            [d], cromwell_config_file=config, cromwell_jar_file=cromwell_jar_file
+        assert  CromwellLocalExecutor(
+            [d], cromwell_configuration=config, cromwell_jar_file=cromwell_jar_file
         ).java_args == f"-Dconfig.file={config}"
 
 
@@ -124,29 +127,29 @@ def test_cromwell_jar():
 
         with setenv({ENV_CROMWELL_JAR: str(jar)}):
             with pytest.raises(FileNotFoundError):
-                CromwellExecutor([d])._cromwell_jar_file
+                 CromwellLocalExecutor([d])._cromwell_jar_file
             with open(jar, "wt") as out:
                 out.write("foo")
-            assert CromwellExecutor([d])._cromwell_jar_file == jar
+            assert  CromwellLocalExecutor([d])._cromwell_jar_file == jar
 
         with setenv({
             ENV_CROMWELL_JAR: None,
             ENV_CLASSPATH: str(d)
         }):
-            assert CromwellExecutor([d])._cromwell_jar_file == jar
+            assert  CromwellLocalExecutor([d])._cromwell_jar_file == jar
 
         with setenv({
             ENV_CROMWELL_JAR: None,
             ENV_CLASSPATH: str(jar)
         }):
-            assert CromwellExecutor([d])._cromwell_jar_file == jar
+            assert  CromwellLocalExecutor([d])._cromwell_jar_file == jar
 
         with setenv({
             ENV_CROMWELL_JAR: None,
             ENV_CLASSPATH: None
         }):
             with pytest.raises(FileNotFoundError):
-                CromwellExecutor([d])._cromwell_jar_file
+                 CromwellLocalExecutor([d])._cromwell_jar_file
 
 
 def test_get_workflow_imports(user_config):
@@ -158,8 +161,7 @@ def test_get_workflow_imports(user_config):
         wdl_dir.mkdir()
         with open(wdl, "wt") as out:
             out.write("foo")
-        executor = CromwellExecutor([wdl_dir],  **cromwell_config)
-        zip_path = executor._get_workflow_imports()
+        zip_path =  CromwellLocalExecutor._get_workflow_imports([wdl_dir])
         assert zip_path.exists()
         with zipfile.ZipFile(zip_path, "r") as import_zip:
             names = import_zip.namelist()
@@ -175,8 +177,7 @@ def test_get_workflow_imports(user_config):
         with open(wdl, "wt") as out:
             out.write("foo")
         imports_file = d / "imports.zip"
-        executor = CromwellExecutor([wdl_dir], **cromwell_config)
-        zip_path = executor._get_workflow_imports(imports_file)
+        zip_path =  CromwellLocalExecutor._get_workflow_imports(import_dirs=[wdl_dir], imports_file=imports_file)
         assert zip_path.exists()
         assert zip_path == imports_file
         with zipfile.ZipFile(zip_path, "r") as import_zip:
@@ -195,8 +196,7 @@ def test_get_workflow_imports(user_config):
         imports_file = d / "imports.zip"
         with open(imports_file, "wt") as out:
             out.write("foo")
-        executor = CromwellExecutor(**cromwell_config)
-        zip_path = executor._get_workflow_imports(imports_file=imports_file)
+        zip_path =  CromwellLocalExecutor._get_workflow_imports(imports_file=imports_file)
         assert zip_path.exists()
         assert zip_path == imports_file
 
@@ -205,7 +205,7 @@ def test_failure_metadata(workflow_data):
     m44 = workflow_data["metadata44.json"]
     with open(m44.path, "rt") as inp:
         m44_dict = json.load(inp)
-    failures = CromwellExecutor._get_failures(m44_dict)
+    failures =  CromwellLocalExecutor._get_failures(m44_dict)
     assert failures.num_failed == 10
     assert failures.failed_task == "contam_testing_set_org_by_volume.org_blast"
     assert failures.failed_task_exit_status == "Unknown"
@@ -217,7 +217,7 @@ def test_failure_metadata(workflow_data):
     m47 = workflow_data["metadata47.json"]
     with open(m47.path, "rt") as inp:
         m47_dict = json.load(inp)
-    failures = CromwellExecutor._get_failures(m47_dict)
+    failures =  CromwellLocalExecutor._get_failures(m47_dict)
     assert failures.num_failed == 10
     assert failures.failed_task == "contam_testing_set_org_by_volume.org_blast"
     assert failures.failed_task_exit_status == "Unknown"
@@ -231,7 +231,7 @@ def test_call_failure_metadata(workflow_data):
     m = workflow_data["metadata_call_failed.json"]
     with open(m.path, "rt") as inp:
         m_dict = json.load(inp)
-    failures = CromwellExecutor._get_failures(m_dict)
+    failures =  CromwellLocalExecutor._get_failures(m_dict)
     assert failures.num_failed == 1
     assert failures.failed_task == "ScatterAt27_14"
     assert failures.failed_task_exit_status == "Unknown"